Put y delete reto 4

+++ mision_tic_back-main/src/main/java/com/example/demo/persistence/repository/CabinRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.persistence.repository;
 import com.example.demo.persistence.entities.Cabin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CabinRepository extends JpaRepository<Cabin, Long> {
+public interface CabinRepository extends JpaRepository<Cabin, Integer> {
 
 }
